+++ test/message.js
@@ -96,6 +96,57 @@ describe('AmiIo.Message', function(){
       done();
     });
 
+    it('parse message with multiple ChanVariableKey correctly', function(){
+      var mes = [
+        "Event: Bridge",
+        "Privilege: call,all",
+        "Timestamp: 1469112499.321389",
+        "Bridgestate: Unlink",
+        "Bridgetype: core",
+        "Channel1: SIP/with-TSE-LIM2-0000a194",
+        "Channel2: Local/5842@from-queue-0000af99;1",
+        "Uniqueid1: 1469111531.131272",
+        "Uniqueid2: 1469111552.131275",
+        "CallerID1: 4959810606",
+        "CallerID2: 5836",
+        "ChanVariable(SIP/with-TSE-LIM2-0000a194): CDR(linkedid)=1469111531.131272",
+        "ChanVariable(SIP/with-TSE-LIM2-0000a194): CDR(dst)=5899",
+        "ChanVariable(Local/5842@from-queue-0000af99;1): CDR(linkedid)=1469111531.131272",
+        "ChanVariable(Local/5842@from-queue-0000af99;1): CDR(dst)=5842"].join('\r\n');
+      message.parse(mes);
+      expect(JSON.stringify(message)).to.deep.equal(JSON.stringify({
+        variables: {},
+        event: 'Bridge',
+        privilege: 'call,all',
+        timestamp: '1469112499.321389',
+        bridgestate: 'Unlink',
+        bridgetype: 'core',
+        channel1: 'SIP/with-TSE-LIM2-0000a194',
+        channel2: 'Local/5842@from-queue-0000af99;1',
+        uniqueid1: '1469111531.131272',
+        uniqueid2: '1469111552.131275',
+        callerid1: '4959810606',
+        callerid2: '5836',
+        'chanvariable(sip/with_tse_lim2_0000a194)': { 'cdr(linkedid)': '1469111531.131272', 'cdr(dst)': '5899' },
+        'chanvariable(local/5842@from_queue_0000af99;1)': { 'cdr(linkedid)': '1469111531.131272', 'cdr(dst)': '5842' },
+        incomingData:
+            [ 'Event: Bridge',
+              'Privilege: call,all',
+              'Timestamp: 1469112499.321389',
+              'Bridgestate: Unlink',
+              'Bridgetype: core',
+              'Channel1: SIP/with-TSE-LIM2-0000a194',
+              'Channel2: Local/5842@from-queue-0000af99;1',
+              'Uniqueid1: 1469111531.131272',
+              'Uniqueid2: 1469111552.131275',
+              'CallerID1: 4959810606',
+              'CallerID2: 5836',
+              'ChanVariable(SIP/with-TSE-LIM2-0000a194): CDR(linkedid)=1469111531.131272',
+              'ChanVariable(SIP/with-TSE-LIM2-0000a194): CDR(dst)=5899',
+              'ChanVariable(Local/5842@from-queue-0000af99;1): CDR(linkedid)=1469111531.131272',
+              'ChanVariable(Local/5842@from-queue-0000af99;1): CDR(dst)=5842' ] }));
+    });
+
   });
 
 });